Allow formatter config to take options/args
This commit introduces `use_format`, which is like the existing `format=` but
optionally accepts arguments/options that will be passed to the formatter's
constructor.

SSHKit's current formatters do not yet accept extra arguments, but this API
lays the groundwork for future formatters or formatter gems like airbrussh that
are configurable and would benefit from a standard API for doing so.