Upgrade to minitest 5
SSHKit was using the `turn` gem, which is no longer maintained. SSHKit was also
pinned to a very old version of the minitest gem (released in 2012).

This commit brings SSHKit up to the current 5.x version of minitest, and
replaces the turn gem with minitest-reporters, which provides similar rspec-like
output with ansi color.

Other changes in minitest 5:

* `MiniTest::Unit::TestCase` is now `Minitest::Test`
* `require 'minitest/autorun'` instead of `MiniTest::Unit.autorun`