Extract ssh channel callbacks into separate methods
Deep nesting of blocks is hard to read and understand.

I have extracted them into several small methods:

* Small individual methods called `handle_EVENT` for each type of event
  that is processed by the library.

* The main setup method `set_handlers` that only sets channel callbacks.
  It does not try to handle the events itself but rather just calls
  the corresponding handler method for each callback.

  In other words, it acts as an event dispatcher.

I believe that this structure is well-maintainable:

* If it is necessary to add a new event handler, one could add a new
  `handle_EVENT` method and a new line to the `set_handlers` method,
  not a whole new block to `execute_command`.

* If it is necessary to change how an event is handled, one could only
  change the corresponding `handle_EVENT` method.

Unfortunately, it was necessary to introduce a whole object with
a method in place of a simple local variable `exit_status`
because only then it was possible to set the status inside
the `handle_exit_status` method.

The new `result` object might hold other values in the future though,
set in a similar way, so that might count as a future-proof thing.