Properly wrap Ruby StandardError w/ add'l context
Prior to this commit, any Ruby StandardError (such as socket connection
errors) would not be rescued or interpreted by SSHKit. In a scenario
where you are running commands on multiple hosts, this meant the error
would be devoid of any context, like on which host the error occurred.

This behavior was actually a mistake: we were intending to rescue and
wrap all Ruby StandardErrors, but the rescue clause was written such
that it only caught `SSHKit::StandardError`, which is something entirely
different.

Fix this by referring to the correct constant, and add unit tests.