Ensure UUID of command is shared when redacted
In logging, the command UUID displayed in the "command start" phase
was sometimes different to that displayed with the output of the
command. This is because the backend can call `.with_redaction` on a
command when logging the start, and this causes the command to be
duplicated. However, if the `uuid` method hadn't been called before
that point, the copy command would generate its own UUID, distinct
from the original command.

This change avoids that by eagerly calculating the UUID when the
command is created, so that any copied objects share the same UUID.