Support older net-sftp API for Ruby 2.0
Old Rubies such as Ruby 2.0 resolve a version of net-sftp that does not
have the `File#size` convenience method. To support these older
versions, use the lower level `File#stat`, which works for versions of
net-sftp all the way back to 2.1.2.

This fixes a `NoMethodError` when SFTP is used via sshkit on Ruby 2.0.