Drop support for Ruby 2.0, 2.1, 2.2, 2.3, and 2.4 (#549)
Already, the latest version of the sshkit gem is no longer installable
on Ruby 2.2 or older. See #548.

Rather than add workarounds to the sshkit code to restore compatibility
with old versions of Ruby, we've decided to officially drop support for
them.

This PR drops support for Ruby 2.0, 2.1, 2.2, 2.3, and 2.4. For
reference, Ruby 2.4 was officially EOL'd 4 years and 9 months ago (31
Mar 2020), so it is not unreasonable for us to drop support in 2025.

If users wish to use the sshkit gem on an old version of Ruby, they can
still do so by installing an older version. For example, 1.22.0 works on
Ruby 2.2.

```
gem install sshkit -v 1.22.0
```