Remove version pin for rainbow dev dependency
The `rainbow` gem is a dependency of rubocop, which we use as a
development dependency on this project. Previously, we were explicitly
declaring rainbow as a dependency so that we could pin it at a specific
version. This was done to ensure the development gems could be installed
on old versions of Ruby in CI (see f57347a).

Now that sshkit no longer supports Ruby < 2.5, this workaround is no
longer needed. This commit removes the explicit rainbow dev dependency.